refactor(storage): move hummock tests to a new crate to break crate cyclic dependency
