+++ Cargo.toml
@@ -24,6 +24,7 @@ members = [
   "src/storage",
   "src/storage/compactor",
   "src/storage/hummock_sdk",
+  "src/storage/hummock_test",
   "src/stream",
   "src/test_runner",
   "src/tests/regress",
