+++ src/storage/hummock_test/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "risingwave_hummock_test"
+version = "0.1.10"
+edition = "2021"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bytes = { version = "1" }
+futures = { version = "0.3" }
+itertools = "0.10"
+rand = "0.8"
+risingwave_common = { path = "../../common" }
+risingwave_hummock_sdk = { path = "../hummock_sdk" }
+risingwave_meta = { path = "../../meta", features = ["test"] }
+risingwave_pb = { path = "../../prost" }
+risingwave_rpc_client = { path = "../../rpc_client" }
+risingwave_storage = { path = "..", features = ["test"] }
+tokio = { version = "=0.2.0-alpha.3", package = "madsim-tokio" }
