+++ src/storage/src/hummock/local_version.rs
@@ -163,7 +163,7 @@ impl PinnedVersion {
         self.version.safe_epoch
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test"))]
     pub fn version(&self) -> HummockVersion {
         self.version.clone()
     }
